// src/divider/types.ts
import { ReactNode } from 'react'

export type DividerProps = {
    /**
     * Тип разделителя, определяющий его ориентацию.
     * В данный момент поддерживается только 'horizontal'.
     */
    type?: 'horizontal' | 'vertical'

    /**
     * Указывает, является ли разделитель прерывистым.
     * Если true, разделитель будет отображаться в виде прерывистой линии.
     */
    dashed?: boolean

    /**
     * Заголовок, отображаемый в центре или сбоку от разделителя.
     */
    title?: string

    /**
     * Ориентация текста заголовка.
     * Опции: 'left' для выравнивания по левому краю,
     * 'center' для центрирования,
     * 'right' для выравнивания по правому краю.
     */
    orientation?: 'left' | 'center' | 'right'

    /**
     * Ширина разделителя.
     */
    width?: string

    /**
     * Высота разделителя.
     */
    height?: string

    /**
     * Отступ между заголовком и линией разделителя.
     */
    gap?: string

    /**
     * Непрозрачность разделителя.
     */
    opacity?: number
    /**
     * Содержимое
     */
    children?: ReactNode | ReactNode[]
}